Fixed player object's location tracking: Player location would always be set to 1.0 so inspecting monster would not work. Also added code author names at function signatures.

+++ src/Player.java
@@ -12,7 +12,7 @@ public class Player {
         return location;
     }
 
-    public void setLocation(int location) {
+    public void setLocation(double location) {
         this.location = location;
     }
 }
